% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_functions.R
\name{BSDE}
\alias{BSDE}
\title{Calculate the BSDE p-value for a SingleCellExperiment object.}
\usage{
BSDE(sce, perm_num = 200, unif_round_unit = 0.2, weight = 1, shrink = FALSE)
}
\arguments{
\item{sce}{A SingleCellExperiment object.
See \code{\link[SingleCellExperiment]{SingleCellExperiment}}.}

\item{perm_num}{The number of permutations for computing Monte Carlo p-value.}

\item{unif_round_unit}{Bin width for empirical histogram (default: 0.2).}

\item{weight}{Weight vector (non-negative, length=number of subjects) for subjects.
Can be used for inverse probability weighting to control for confounding.
Default: 1 for equal weights.}

\item{shrink}{If \code{TRUE}, the range with zero count will be removed to
cope with sparsity and speed up the computation. Use \code{TRUE} if
extremely high expressions are present. Default: \code{FALSE}.}
}
\value{
a list of p-values
}
\description{
Calculate the BSDE p-value for a SingleCellExperiment object.
}
