% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_functions.R
\name{calc_zinb_param}
\alias{calc_zinb_param}
\title{this function returns the parameters for changing mean and variance of the zero-inflated negative binomial(ZINB) distribution}
\usage{
calc_zinb_param(mu, theta, drop = 0, r_m = 1, r_v = 1)
}
\arguments{
\item{mu}{the original parameter mu of ZINB distribtuion}

\item{theta}{the original parameter theta of ZINB distribtuion, the distribution will be more close to zero-inflated poission with larger parameter size.}

\item{drop}{the original parameter dropout_rate of of ZINB distribution}

\item{r_m}{the targeted fold change of mean}

\item{v_m}{the targeted fold change of variance}
}
\value{
two number for parameter mu and theta of modified ZINB distribution
}
\description{
this function returns the parameters for changing mean and variance of the zero-inflated negative binomial(ZINB) distribution
}
\note{
For numerical meaningful calculation, is is required that r_m/r_v <1+mu/theta
}
