% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{de.dp}
\alias{de.dp}
\title{indicator vector (1 for differential expression in proportion).}
\format{
An object of class \code{numeric} of length 30.
}
\usage{
de.dp
}
\description{
indicator vector (1 for differential expression in proportion).
}
\keyword{datasets}
