% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{de.mean}
\alias{de.mean}
\title{indicator vector (1 for differential expression in mean)}
\format{
An object of class \code{numeric} of length 30.
}
\usage{
de.mean
}
\description{
indicator vector (1 for differential expression in mean)
}
\keyword{datasets}
