% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{de.mult}
\alias{de.mult}
\title{indicator vector (1 for differential expression in multimodality)}
\format{
An object of class \code{numeric} of length 30.
}
\usage{
de.mult
}
\description{
indicator vector (1 for differential expression in multimodality)
}
\keyword{datasets}
