% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{de.var}
\alias{de.var}
\title{indicator vector (1 for differential expression in variance)}
\format{
An object of class \code{numeric} of length 30.
}
\usage{
de.var
}
\description{
indicator vector (1 for differential expression in variance)
}
\keyword{datasets}
