% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce}
\alias{sce}
\title{\code{SingleCellExperiment::SingleCellExperiment} object}
\format{
An object of class \code{SingleCellExperiment} with 30 rows and 800 columns.
}
\usage{
sce
}
\description{
\code{SingleCellExperiment::SingleCellExperiment} object
}
\keyword{datasets}
