% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_matrix}
\alias{sim_matrix}
\title{count data (row: gene, column: cell)}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 30 rows and 800 columns.
}
\usage{
sim_matrix
}
\description{
count data (row: gene, column: cell)
}
\keyword{datasets}
