% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_matrix_bulk}
\alias{sim_matrix_bulk}
\title{matrix for bulk RNA data (gene x individual)}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 30 rows and 40 columns.
}
\usage{
sim_matrix_bulk
}
\description{
matrix for bulk RNA data (gene x individual)
}
\keyword{datasets}
