% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_functions.R
\name{simu_base_param}
\alias{simu_base_param}
\title{This function works for the basic parameter generation of ZINB models from the given reference data.}
\usage{
simu_base_param(
  t_mean,
  t_disp,
  t_drop,
  t_meta,
  nTotal = 30,
  nall = 40,
  RIN_adj = TRUE
)
}
\arguments{
\item{t_mean}{a genexcell matrix, represent the mu parameter of ZINB models from reference data. DCA can generate it.}

\item{t_disp}{a genexcell matrix, represent the overdispersion parameter of ZINB models from reference data. DCA can generate it.}

\item{t_drop}{a genexcell matrix, represent the dropout_rate parameter of ZINB models from reference data. DCA can generate it.}

\item{nTotal}{numbers of subjects needs to be simulated}
}
\value{
a list includes two objects sample_ctrl and RNA.simu

sample_ctrl is a gene x individual x param array, which gives the parameter "mean", "dispersion", and "dropout" for simulate a ZINB distribution.
It also gives the residual standard deviations of means.
}
\description{
This function works for the basic parameter generation of ZINB models from the given reference data.
}
\references{
Eraslan, G., Simon, L. M., Mircea, M., Mueller, N. S., & Theis, F. J. (2019). Single-cell RNA-seq denoising using a deep count autoencoder. Nature communications, 10(1), 1-14.
}
