% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_functions.R
\name{WaBarycenter2}
\alias{WaBarycenter2}
\title{This function is modified for 1-dimensional data from WaBarycenter function of R package 'Barycenter', and used for the option "R" of merge_method at function cal_w2_pval
A list of matrices satisfying the prerequisites described above.}
\usage{
WaBarycenter2(
  images,
  maxIter = 10,
  lambda = FALSE,
  costm = FALSE,
  lambda_fold = 1
)
}
\arguments{
\item{images}{The input vector, which could be the log-transformed and normalized expression counts.}

\item{maxIter}{Maximum number of iterations.}

\item{lambda}{Non-negative regularization parameter (for large lambda the regularized Barycenter is close to its true counterpart). If FALSE the algorithm uses a lambda depending on costm.}

\item{costm}{A matrix of pairwise distances between the locations. If FALSE the algorithm uses the usual euclidean distance matrix on a [0,1]x[0,1] equidistant pixel grid.}

\item{lambda_fold}{A modification for the default lambda adjustment.}
}
\value{
Value The Barycenter of the input, represented by a  vector.
}
\description{
This function is modified for 1-dimensional data from WaBarycenter function of R package 'Barycenter', and used for the option "R" of merge_method at function cal_w2_pval
A list of matrices satisfying the prerequisites described above.
}
\references{
Cuturi, M.: Fast Computation of Wasserstein Barycenters, Proceedings of the International Conference on Machine Learning, Beijing, China, 2014
}
