% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_functions.R
\name{cal_w2_bulk_pval}
\alias{cal_w2_bulk_pval}
\title{This function calculates the w2_pval based on BSDE for a given gene, based on bulk RNAseq.
The expression are represented as one numerical value per gene, per individual.}
\usage{
cal_w2_bulk_pval(
  cur_ind_count,
  cur_individual,
  phenotype,
  perm_num = 500,
  unif_round_unit = 0.5
)
}
\arguments{
\item{count_per_gene:}{len x 1 matrix,length=ind num,  gene count}

\item{meta_individual:}{len x 1 vector, length=ind num, indicator labels (contains 0,1)}

\item{perm_num:ingeter, }{permutation time}
}
\value{
a list contains:
pval a p-value, from the based on Monte Carlo Permutation Procedure.
case_distr a vectors contains the density of case subjects distribution of Barycenter
ctrl_distr a vectors contains the density of ctrl subjects distribution of Barycenter
}
\description{
This function calculates the w2_pval based on BSDE for a given gene, based on bulk RNAseq.
The expression are represented as one numerical value per gene, per individual.
}
