% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_functions.R
\name{cal_w2_pval}
\alias{cal_w2_pval}
\title{This function calculates the BSDE Barycenter p-value of a given gene.}
\usage{
cal_w2_pval(
  count_per_gene,
  meta_individual,
  meta_phenotype,
  perm_num = 200,
  unif_round_unit = 0.2,
  weight = 1,
  shrink = FALSE
)
}
\arguments{
\item{count_per_gene}{Gene count. Vector of length n (n is number of cells).}

\item{meta_individual}{Label of individuals (length n).}

\item{meta_phenotype}{0/1 vector of length n (0=control, 1=case).}

\item{perm_num}{Number of random permutations for computing p-value.}

\item{unif_round_unit}{Bin width for empirical histogram (default: 0.2).}

\item{weight}{Weight vector (non-negative, length=number of subjects) for subjects.
Can be used for inverse probability weighting to control for confounding.
Default: 1 for equal weights.}

\item{shrink}{If \code{TRUE}, the range with zero count will be removed to
cope with sparsity and speed up the computation. Use \code{TRUE} if
extremely high expressions are present. Default: \code{FALSE}.}
}
\value{
a list contains:
\itemize{
\item pval: p-value
\item case_bc_ob: density of the aggregated Barycenter distribution of the cases
\item ctrl_bc_ob: density of the aggregated Barycenter distribution of the controls
}
}
\description{
This function calculates the BSDE Barycenter p-value of a given gene.
}
