% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_functions.R
\name{calc_multimod_param}
\alias{calc_multimod_param}
\title{For a given mixture \eqn{0.5 ZINB(\mu(1-t), \theta, z) + 0.5 ZINB(\mu(1+t), \theta, z)},
return \eqn{ZINB(\mu, \theta', z)} such that both mean and variance match.}
\usage{
calc_multimod_param(size, t = 0.5)
}
\arguments{
\item{size}{parameter \eqn{\theta} of the ZINB model.}

\item{t}{Parameter \eqn{t} (between 0 and 1).}
}
\value{
Parameter \eqn{\theta'} of the output ZINB model.
}
\description{
For a given mixture \eqn{0.5 ZINB(\mu(1-t), \theta, z) + 0.5 ZINB(\mu(1+t), \theta, z)},
return \eqn{ZINB(\mu, \theta', z)} such that both mean and variance match.
}
