% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_functions.R
\name{calc_zinb_param}
\alias{calc_zinb_param}
\title{Given a ZINB(\eqn{\mu}, \eqn{\theta}, \eqn{z}), returns ZINB(\eqn{\mu'}, \eqn{\theta'}, \eqn{z})
such that the mean changes by a factor of \eqn{r_m} and the variance changes
by a factor of \eqn{v_m}.}
\usage{
calc_zinb_param(mu, theta, drop = 0, r_m = 1, r_v = 1)
}
\arguments{
\item{mu}{parameter \eqn{\mu} of ZINB distribution.}

\item{theta}{parameter \eqn{\theta} of ZINB distribution.}

\item{drop}{dropout rate \eqn{z} of ZINB distribution.}

\item{r_m}{targeted fold change of mean}

\item{r_v}{targeted fold change of variance}
}
\value{
\eqn{\mu'} and \eqn{\theta'} of the modified ZINB distribution
}
\description{
Given a ZINB(\eqn{\mu}, \eqn{\theta}, \eqn{z}), returns ZINB(\eqn{\mu'}, \eqn{\theta'}, \eqn{z})
such that the mean changes by a factor of \eqn{r_m} and the variance changes
by a factor of \eqn{v_m}.
}
\note{
It is required that \eqn{r_m / r_v < 1 + \mu / \theta}.
}
\examples{
calc_zinb_param(2,3,r_m=1.3)
calc_zinb_param(2,3,r_v=1.5)

}
