% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_functions.R
\name{simu_base_param}
\alias{simu_base_param}
\title{Given parameters of a reference dataset, calculate the ZINB parameters for
simulating a new dataset of a different size.}
\usage{
simu_base_param(
  t_mean,
  t_disp,
  t_drop,
  t_meta,
  nTotal = 30,
  nall = 40,
  RIN_adj = FALSE
)
}
\arguments{
\item{t_mean}{mean parameter matrix \eqn{\mu} of the reference dataset (gene x cell).}

\item{t_disp}{dispersion parameter matrix \eqn{\theta} of the reference dataset (gene x cell).}

\item{t_drop}{dropout parameter matrix \eqn{z} of the reference dataset (gene x cell).}

\item{t_meta}{Data frame (nrow = cell) of meta information.
It must contain a column \code{individual} recording from which subject each cell is collected.
If \code{RIN_adj=TRUE}, then it must also contain column \code{RIN} for RNA integrity number.}

\item{nTotal}{numbers of genes to simulate}

\item{nall}{numbers of individuals to simulate}

\item{RIN_adj}{If \code{TRUE}, then \code{t_meta} must contain column \code{RIN} for RNA integrity number.
(default: FALSE)}
}
\value{
a list containing:
\itemize{
\item \code{sample_ctrl}: gene x individual x param array,
which gives the parameter \eqn{\mu}, \eqn{\theta} and \eqn{z} of ZINB for simulation.
\item \code{RNA.simu}: residual standard deviations of means.
}
}
\description{
The basic parameters of ZINB(\eqn{\mu}, \eqn{\theta}, \eqn{z}) from the reference dataset can be
estimated with DCA (Deep Count Autoencoder).
}
\references{
Eraslan, G., Simon, L. M., Mircea, M., Mueller, N. S., & Theis, F. J. (2019). Single-cell RNA-seq denoising using a deep count Autoencoder. Nature Communications, 10(1), 1-14.
}
